# Conflicts:
#	dev/Code/CryEngine/CryCommon/VertexFormats.h
#	dev/Code/CryEngine/RenderDll/Common/Shaders/Vertex.h
#	dev/Code/CryEngine/RenderDll/Common/tests/Shaders/VertexTests.cpp

+++ dev/Code/Sandbox/Plugins/CryDesigner/Tools/Edit/OffsetTool.cpp
@@ -261,7 +261,7 @@ void OffsetTool::Display(DisplayContext& dc)
 
     if (m_pOffsetedPolygon)
     {
-        int oldThickness = dc.GetLineWidth();
+        float oldThickness = dc.GetLineWidth();
         dc.SetLineWidth(CD::kLineThickness);
 
         dc.SetColor(CD::kResizedPolygonColor);
